Search flask logic improved, portfolio buttons and db added
Flask logic for returning the search results is now simplified. Function seperated to return list of dictionary of companies to display, called everytime a search is made or a button is pressed. Portfolio database added, search page only adds stock info, quantities are to be added to db from the portfolio page